fix(file): Restructure files and add index.ts in src/features/company-register
