+++ src/features/company-register/index.ts
@@ -1,0 +1,1 @@
+export { CompanyRegisterForm } from "./components/CompanyRegisterForm";
